+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.17.0] 2021-07-28
+
+### Added
+- regexp entities now can be treated without diacritics and with whole words matching
+
+### Changed
+- published entity matching methods on CustomEntityDetectionModel
+
+
 ## [3.16.1] 2021-06-21
 
 ### Fixed